Create Map
It will initialize a google maps instance, and after that it'll add
markers if the markersCoordinates parameter is passed.
